requirements.txt
django-registration
simplejson
django
dj-database-url==0.4.1
django-bootstrap3
Pillow
djangorestframework
markdown
django-filter
gunicorn==19.6.0
django-geoposition
whitenoise==2.0.6
django-versatileimagefield
django-ckeditor
python-dateutil
django-admin-interface==0.1.4
django-autofixture==0.12.0
django-colorfield==0.1.10
django-crispy-forms==1.6.0
django-daterange-filter==1.3.0
django-extensions==1.6.1
django-filters-mixin==0.1.3
django-flat-theme==1.1.2
django-grappelli==2.8.1
django-import-export==0.4.3
django-simple-history==1.8.1
django-vanilla-views==1.0.4
psutil==4.2.0
pyasn1==0.1.9
Pygments==2.1.3
pyOpenSSL==16.1.0